+++ src/main/java/org/scrumbucket/crawler4neo/services/CrawlerService.java
@@ -1,4 +1,4 @@
-package org.scrumbucket.crawler4neo.crawler;
+package org.scrumbucket.crawler4neo.services;
 
 
 import org.scrumbucket.crawler4neo.data.Page;
