+++ src/main/java/org/scrumbucket/crawler4neo/services/GrabManager.java
@@ -1,4 +1,4 @@
-package org.scrumbucket.crawler4neo.crawler;
+package org.scrumbucket.crawler4neo.services;
 
 import org.apache.commons.lang3.StringUtils;
